Fix InMemoryStreamLog

+++ test/src/test/java/org/corfudb/integration/AbstractIT.java
@@ -144,7 +144,7 @@ public class AbstractIT extends AbstractCorfuTest {
 
             ProcessBuilder builder = new ProcessBuilder();
             for (Long pid : pidList) {
-                builder.command("sh", "-c", KILL_COMMAND + pid.longValue());
+                builder.command("sh", "-c", KILL_COMMAND + pid);
                 Process p = builder.start();
                 p.waitFor();
             }
